Feature(app): исправил ошибки сборки
-Изменил все значения типа Double и Int на тип BigDecimal
-Написал функцию-расширение для суммирования по BigDecimal

TICKER: ANDROID-149

+++ app/src/main/java/com/infinity_coder/divcalendar/domain/models/EditPaymentParams.kt
@@ -1,8 +1,10 @@
 package com.infinity_coder.divcalendar.domain.models
 
+import java.math.BigDecimal
+
 data class EditPaymentParams(
     val portfolioId: Long,
     val isin: String,
     val date: String,
-    val count: Int
+    val count: BigDecimal
 )
